requirements.txt
numpy>=2.1.3
pandas>=2.2.3
matplotlib>=3.10.0
torch>=2.6.0+cu118
PyYAML>=6.0.1
openpyxl>=3.1.2
xlsxwriter>=3.2.0